Wireshark Config: Signal PDU bugfix
Signal PDUs over SOME/IP had an issue with request/response methods,
getters, and setters. The generator generated for both directions
the same PDU-ID, which leads to clashed.

This issues was not severe, since for getters and setters the
generated config was working, while Signal PDUs over generic
request/response methods has no real use case.

In addition, the code was refactored as well, since the generation
of PDU-IDs had to be changed. A field now gets exactly one PDU.